·Created file constants. ·Created schemes folder validators in middlewares.
