+++ app/constants.js
@@ -1,0 +1,2 @@
+exports.regDomainWolox = /^(([^<>()[\]\\,;:\s@”]+((@wolox)+\.(co))))$/;
+exports.regPasswordAlphaNumeric = /^(?=.*\d)(?=.*\D)/;
