+++ app/routes.js
@@ -1,8 +1,11 @@
 // const controller = require('./controllers/controller');
+
+const { schemaSignUp } = require('./middlewares/schemas/signs_up');
+
+const { validationSchema, encryptPassword } = require('./middlewares/verify_signs_up');
+
 const { healthCheck } = require('./controllers/healthCheck');
 const { createUser } = require('./controllers/users');
-
-const { schemaSignUp, validationSchema, encryptPassword } = require('./middlewares/users/verify_signs_up');
 
 exports.init = app => {
   app.get('/health', healthCheck);
